Examples/PyChinook/pyproject.toml
[project]
name = "pychinook"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "mssql-python==1.0.0a0",
    "pyodbc==5.2.0",
    "pytest==8.4.1",
    "python-dotenv==1.1.1",
    "sqlalchemy>=2.0.43",
]
